Updated the code after integration testing with all loaders
